Make memory mapping more accurate
The memory buffers will be the correct size (e.g. 2K for system and
PPU RAM) and the read_mem/write_mem functions will correctly map
the right addresses to the right place and hande mirroring.

A cartridge module is added to do the part of the memory mapping
that is handled by the cartridge/mapper on a real system.